refactor: 인기 콘텐츠 페이지에 쿼리 서스펜스 경계 추가 및 인기 카드 컴포넌트 리팩토링
